@@ -1,10 +1,12 @@
-// lib/presentation/screens/splash/splash_screen.dart
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import '../../../core/constants/app_routes.dart';
+import '../../../core/theme/app_theme.dart';
 import '../../providers/auth_provider.dart';
-import '../../../core/constants/app_constants.dart';
 
+/// Splash screen with elegant animations
+/// Shows app branding and handles initial navigation
 class SplashScreen extends ConsumerStatefulWidget {
   const SplashScreen({super.key});
 
@@ -13,123 +15,293 @@ class SplashScreen extends ConsumerStatefulWidget {
 }
 
 class _SplashScreenState extends ConsumerState<SplashScreen>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _animationController;
+    with TickerProviderStateMixin {
+  late AnimationController _logoController;
+  late AnimationController _textController;
+  late AnimationController _fadeController;
+
+  late Animation<double> _logoScaleAnimation;
+  late Animation<double> _logoRotationAnimation;
+  late Animation<double> _textSlideAnimation;
+  late Animation<double> _textOpacityAnimation;
   late Animation<double> _fadeAnimation;
-  late Animation<double> _scaleAnimation;
 
   @override
   void initState() {
     super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(seconds: 2),
+    _initializeAnimations();
+    _startAnimations();
+    _checkAuthAndNavigate();
+  }
+
+  void _initializeAnimations() {
+    // Logo animations
+    _logoController = AnimationController(
+      duration: const Duration(milliseconds: 1500),
       vsync: this,
     );
 
-    _fadeAnimation = Tween<double>(
-      begin: 0.0,
-      end: 1.0,
-    ).animate(CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.easeIn,
-    ));
-
-    _scaleAnimation = Tween<double>(
-      begin: 0.5,
-      end: 1.0,
-    ).animate(CurvedAnimation(
-      parent: _animationController,
+    _logoScaleAnimation = CurvedAnimation(
+      parent: _logoController,
       curve: Curves.elasticOut,
+    );
+
+    _logoRotationAnimation = Tween<double>(
+      begin: 0,
+      end: 2 * 3.14159,
+    ).animate(CurvedAnimation(
+      parent: _logoController,
+      curve: Curves.easeInOut,
     ));
 
-    _animationController.forward();
-    _navigateToNextScreen();
+    // Text animations
+    _textController = AnimationController(
+      duration: const Duration(milliseconds: 1200),
+      vsync: this,
+    );
+
+    _textSlideAnimation = Tween<double>(
+      begin: 30,
+      end: 0,
+    ).animate(CurvedAnimation(
+      parent: _textController,
+      curve: Curves.easeOutCubic,
+    ));
+
+    _textOpacityAnimation = CurvedAnimation(
+      parent: _textController,
+      curve: Curves.easeIn,
+    );
+
+    // Fade animation
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 500),
+      vsync: this,
+    );
+
+    _fadeAnimation = CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeIn,
+    );
+  }
+
+  void _startAnimations() async {
+    await Future.delayed(const Duration(milliseconds: 300));
+    _logoController.forward();
+
+    await Future.delayed(const Duration(milliseconds: 600));
+    _textController.forward();
+  }
+
+  void _checkAuthAndNavigate() async {
+    // Wait for animations to complete
+    await Future.delayed(const Duration(seconds: 3));
+
+    // Start fade out
+    await _fadeController.forward();
+
+    if (!mounted) return;
+
+    // Check authentication status
+    final authState = ref.read(authStateProvider);
+
+    if (authState.isAuthenticated) {
+      if (authState.isProfileComplete) {
+        context.go(AppRoutes.home);
+      } else {
+        context.go(AppRoutes.profileSetup);
+      }
+    } else {
+      // Check if it's first time user
+      final isFirstTime = await ref.read(authServiceProvider).isFirstTimeUser();
+
+      if (isFirstTime) {
+        context.go(AppRoutes.onboarding);
+      } else {
+        context.go(AppRoutes.signIn);
+      }
+    }
   }
 
   @override
   void dispose() {
-    _animationController.dispose();
+    _logoController.dispose();
+    _textController.dispose();
+    _fadeController.dispose();
     super.dispose();
-  }
-
-  void _navigateToNextScreen() async {
-    await Future.delayed(const Duration(seconds: 3));
-
-    if (mounted) {
-      final authState = ref.read(authStateProvider);
-      authState.when(
-        loading: () => context.go('/auth/login'),
-        error: (error, stackTrace) => context.go('/auth/login'),
-        data: (user) {
-          if (user != null) {
-            context.go('/home');
-          } else {
-            context.go('/onboarding');
-          }
-        },
-      );
-    }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).primaryColor,
-      body: Center(
-        child: AnimatedBuilder(
-          animation: _animationController,
-          builder: (context, child) {
-            return FadeTransition(
-              opacity: _fadeAnimation,
-              child: ScaleTransition(
-                scale: _scaleAnimation,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Container(
-                      width: 120,
-                      height: 120,
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                        borderRadius: BorderRadius.circular(30),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withOpacity(0.3),
-                            blurRadius: 20,
-                            spreadRadius: 5,
+      body: FadeTransition(
+        opacity: Tween<double>(begin: 1, end: 0).animate(_fadeAnimation),
+        child: Container(
+          decoration: const BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                Color(0xFFF8F5F2), // Warm white
+                Color(0xFFE8E2DC), // Light beige
+              ],
+            ),
+          ),
+          child: SafeArea(
+            child: Center(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  // Animated logo
+                  AnimatedBuilder(
+                    animation: _logoController,
+                    builder: (context, child) {
+                      return Transform.scale(
+                        scale: _logoScaleAnimation.value,
+                        child: Transform.rotate(
+                          angle: _logoRotationAnimation.value,
+                          child: Container(
+                            width: 120,
+                            height: 120,
+                            decoration: BoxDecoration(
+                              shape: BoxShape.circle,
+                              gradient: const LinearGradient(
+                                begin: Alignment.topLeft,
+                                end: Alignment.bottomRight,
+                                colors: [
+                                  AppTheme.primaryPurple,
+                                  Color(0xFF8B7FE8),
+                                ],
+                              ),
+                              boxShadow: [
+                                BoxShadow(
+                                  color: AppTheme.primaryPurple.withOpacity(0.3),
+                                  blurRadius: 30,
+                                  offset: const Offset(0, 15),
+                                ),
+                              ],
+                            ),
+                            child: const Icon(
+                              Icons.spa_rounded,
+                              size: 60,
+                              color: Colors.white,
+                            ),
                           ),
-                        ],
-                      ),
-                      child: const Icon(
-                        Icons.face,
-                        size: 60,
-                        color: Color(0xFF6C5CE7),
-                      ),
-                    ),
-                    const SizedBox(height: 24),
-                    const Text(
-                      'SkinSense',
-                      style: TextStyle(
-                        fontSize: 32,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    const Text(
-                      'AI-Powered Skin Analysis',
-                      style: TextStyle(
-                        fontSize: 16,
-                        color: Colors.white70,
-                      ),
-                    ),
-                  ],
-                ),
+                        ),
+                      );
+                    },
+                  ),
+                  const SizedBox(height: 40),
+
+                  // Animated app name
+                  AnimatedBuilder(
+                    animation: _textController,
+                    builder: (context, child) {
+                      return Transform.translate(
+                        offset: Offset(0, _textSlideAnimation.value),
+                        child: Opacity(
+                          opacity: _textOpacityAnimation.value,
+                          child: Column(
+                            children: [
+                              Text(
+                                'ALPHA',
+                                style: Theme.of(context).textTheme.displaySmall?.copyWith(
+                                  color: AppTheme.darkGray,
+                                  fontWeight: FontWeight.w300,
+                                  letterSpacing: 8,
+                                ),
+                              ),
+                              Text(
+                                'SKINCARE',
+                                style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                                  color: AppTheme.primaryPurple,
+                                  fontWeight: FontWeight.w600,
+                                  letterSpacing: 4,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+                  const SizedBox(height: 20),
+
+                  // Animated tagline
+                  AnimatedBuilder(
+                    animation: _textController,
+                    builder: (context, child) {
+                      return Opacity(
+                        opacity: _textOpacityAnimation.value,
+                        child: Text(
+                          'Love Your Skin Better',
+                          style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                            color: AppTheme.mediumGray,
+                            fontStyle: FontStyle.italic,
+                            letterSpacing: 1,
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+                  const SizedBox(height: 80),
+
+                  // Loading indicator
+                  AnimatedBuilder(
+                    animation: _textController,
+                    builder: (context, child) {
+                      return Opacity(
+                        opacity: _textOpacityAnimation.value,
+                        child: Container(
+                          width: 100,
+                          height: 3,
+                          decoration: BoxDecoration(
+                            color: AppTheme.lightBeige,
+                            borderRadius: BorderRadius.circular(2),
+                          ),
+                          child: LayoutBuilder(
+                            builder: (context, constraints) {
+                              return Stack(
+                                children: [
+                                  AnimatedContainer(
+                                    duration: const Duration(seconds: 2),
+                                    width: constraints.maxWidth,
+                                    height: 3,
+                                    decoration: BoxDecoration(
+                                      gradient: const LinearGradient(
+                                        colors: [
+                                          AppTheme.primaryPurple,
+                                          Color(0xFF8B7FE8),
+                                        ],
+                                      ),
+                                      borderRadius: BorderRadius.circular(2),
+                                    ),
+                                  ),
+                                ],
+                              );
+                            },
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+                ],
               ),
-            );
-          },
+            ),
+          ),
         ),
       ),
     );
   }
+}
+
+// Temporary theme references
+class AppTheme {
+  static const Color primaryPurple = Color(0xFF6C5CE7);
+  static const Color secondaryGreen = Color(0xFF00B894);
+  static const Color warmWhite = Color(0xFFF8F5F2);
+  static const Color lightBeige = Color(0xFFE8E2DC);
+  static const Color darkGray = Color(0xFF2D2D2D);
+  static const Color mediumGray = Color(0xFF666666);
+  static const Color lightGray = Color(0xFF999999);
 }
